"cd", "main" and "man"
Added cd command as a built-in function of the shell. I did some small
changes to the code of cd (added a header file and "extern"-ed the
function). I also added the command that Jon made (he was not sure how
to use push so he asked me to move its file here).