Changes in project directory, main.c and defines.h
I have given a little bit of order to the structure of the code of
main.c, now it is a little bit easier to read. I have also added
documentation to the auxiliary functions of this code.
I have decided to change the working directory of the shell
when it starts, now we always start at ./data/home.
I have made changes to defines.h, now the libraries that we most
often use are located there and can be included by just using
'#include "defines.h"', I have also added a externed variable
to store the path to the most important directories of the project.
I made a little change in the script "createGameDirectory" as to
make it generate the .cd of home (it was not getting generated). I
also changed it from directory and moved to a special directory
called "scripts" in /bin/.
Finally, I created the directory "docs" inside /data/ to store the
information that man.c is going to print. I also decided to create
a little file to store the welcoming message of the shell.