Small changes, mostly documents.
Small revision and correction of documents.
Changed the images and the .svg
Changed the way the named pipe is created in fifo_write
Deleted the "logout" and only left the "exit" command.
Added some includes to ls.c
Added two doc info for the man.
deleteGameDirectory doesn't print anything to stdout now.